Move models into appropriate file
The models in k8s/models/apiextensions_v1.py are only used by CRD.
This moves them into
k8s/models/apiextensions_v1_custom_resource_definition.py
and deletes k8s/models/apiextensions_v1.py
